+++ infrastructure/server/oidc_clients_spar.tf
@@ -18,7 +18,7 @@ resource "aws_cognito_user_pool_client" "dev_spar_oidc_client" {
   id_token_validity                             = "60"
   name                                          = "spar_dev"
   prevent_user_existence_errors                 = "ENABLED"
-  read_attributes                               = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name"])}"
+  read_attributes                               = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
   refresh_token_validity                        = "24"
   supported_identity_providers                  = [
     "${aws_cognito_identity_provider.dev_idir_oidc_provider.provider_name}",
@@ -32,7 +32,7 @@ resource "aws_cognito_user_pool_client" "dev_spar_oidc_client" {
   }
 
   user_pool_id     = aws_cognito_user_pool.fam_user_pool.id
-  write_attributes = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name"])}"
+  write_attributes = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
 }
 
 resource "aws_cognito_user_pool_client" "test_spar_oidc_client" {
@@ -55,7 +55,7 @@ resource "aws_cognito_user_pool_client" "test_spar_oidc_client" {
   id_token_validity                             = "60"
   name                                          = "spar_test"
   prevent_user_existence_errors                 = "ENABLED"
-  read_attributes                               = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name"])}"
+  read_attributes                               = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
   refresh_token_validity                        = "24"
   supported_identity_providers                  = [
     "${aws_cognito_identity_provider.test_idir_oidc_provider.provider_name}",
@@ -69,7 +69,7 @@ resource "aws_cognito_user_pool_client" "test_spar_oidc_client" {
   }
 
   user_pool_id     = aws_cognito_user_pool.fam_user_pool.id
-  write_attributes = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name"])}"
+  write_attributes = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
 }
 
 resource "aws_cognito_user_pool_client" "prod_spar_oidc_client" {
@@ -89,7 +89,7 @@ resource "aws_cognito_user_pool_client" "prod_spar_oidc_client" {
   id_token_validity                             = "60"
   name                                          = "spar_prod"
   prevent_user_existence_errors                 = "ENABLED"
-  read_attributes                               = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name"])}"
+  read_attributes                               = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
   refresh_token_validity                        = "24"
   supported_identity_providers                  = [
     "${aws_cognito_identity_provider.prod_idir_oidc_provider.provider_name}",
@@ -103,5 +103,5 @@ resource "aws_cognito_user_pool_client" "prod_spar_oidc_client" {
   }
 
   user_pool_id     = aws_cognito_user_pool.fam_user_pool.id
-  write_attributes = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name"])}"
+  write_attributes = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
 }